Optimize dumping objects by code generating functions to dump objects.
After profiling some of our services running Marshmallow, we noticed that it was
consuming a ton of CPU.  This approach is similar to that of namedtuple in that
we generate python code for dumping objects.  We then fall back to the original,
slower path should anything go wrong.

In the example provided in performance/benchmark.py we see a 16x speedup.  For
some of our internal schemas, this is closer to 20x.

```
Benchmark Result:
	Original Time: 2382.36 usec/dump
	Optimized Time: 145.83 usec/dump
	Speed up: 16.34x
```