Allow Number fields to prohibit boolean values.
After a bunch of research, I decided to _explicitly_ special case booleans as opposed
to adding a flag for only allowing numbers to be serialized.  I came across code in
our repositories that, for example, overload `__int__` or `__float__` that would expect
to be able to be serialized.  I agree, however, that allowing booleans to be serialized
as numbers is likely incorrect.

This fixes https://github.com/marshmallow-code/marshmallow/issues/623