Replace set_value(allow_dotted=False) with setitem
Rather than adding the feature for utils.set_value() to do this, just
do dict key-value assignment. However, this results in a difficult to
read expression: `typing.cast(typing.Dict, ret)[key]`, when all that
is really *meant* is `ret[key]`.

Because mypy assigns types to variables and does not allow those types
to be reassigned or altered -- excepting type narrowing on unions --
the easy way out is to use different names for the `many=True/False`
branches of `_deserialize`. The branches now use `ret_l` (l for list)
and `ret_d` (d for dict) and there is no shared variable named `ret`.
This removes casts of `ret` to a dict, and results in
`ret_d[key] = value` in this case, which is the simplest/most readable
form.