Allow `fields.Nested` to consume a dict
To make it easier to leverage `Schema.from_dict`, allow `Nested` to
consume a dict mapping names to fields. This is internally passed to
`Schema.from_dict`.
That is, what previously would be written

    Nested(Schema.from_dict({"foo": String()}))

can now be written

    Nested({"foo": String()})

This relates to the original use-case for `Schema.from_dict`: easier
lightweight schema generation.

The conversion of dicts also supports a callable which returns a dict.
That is, the following is also supported:

    Nested(lambda: {"foo": String()})