Add messages_dict property to ValidationError
'messages_dict' is an alias for 'messages' which checks if the value
is a dict and raises a TypeError if it is not.

This "type asserting property" leads to easier usage with type
checkers. In many cases, the 'messages' attribute is known to be a
dict, but is annotated as a union type, leading to false negatives
from type checkers. Using 'messages_dict', one can write

    x: dict = err.messages_dict

rather than

    x: dict = typing.cast(dict, err.messages)

To validate this behavior at type checking time and runtime, two tests
are added:

1. A pytest test which checks that the desired TypeError is raised if
   the value is not a dict

2. A new file to act as a "mypy test case", using
   `mypy --warn-unused-ignores` to check that things are annotated
   correctly

(1) is like existing tests, but (2) requires a little bit of new
infrastrucutre. A new dir for the tests, which is listed in pytest
norecursedirs, and a new tox env (listed in CI as well) which runs
`mypy` with the desirable flags.

resolves #1442

Ignore mypy_test_cases in pre-commit. mypy checking running without
`marshmallow` installed won't check these files correctly. They're
explicitly there to be tested under tox, so exclude and add a
comment.