Validate values passed for 'unknown'
There are three places that a user can pass a value for 'unknown':
- Meta in a Schema class definition
- Schema.__init__
- Schema.load

For each of these, call an internal check to confirm that the value is
valid. In cases where `unknown or self.unknown` is used to support
`None`, do the validation after the `or`.

A new test case covers these three locations and confirms that passing
an unknown (hah!) value causes a ValueError.

closes #1721, #1732